requirements.txt
altgraph==0.15
construct==2.8.22
docopt==0.6.2
future==0.16.0
macholib==1.9
numpy==1.13.3
olefile==0.44
pefile==2017.11.5
Pillow==5.0.0
PyInstaller==3.3.1
